Fix - BookListScreen: Tap on empty area of book item triggers action
Issue : When a user taps occurs on an area of the book item that is not
text, it did'nt trigger the interaction handler. This was because the
text did not span the scrollview containers and the taps were captured
by these containers on which no listeners were set.

Fix - Make the scrollview fill the views they contain

Also
Updated the names of strign resources and view Ids to follow
nomenclature of the project.