Fix-Do not forcefully show the Main NavBar on detaching the fragment
Bug:
Navigate to ChapterScreen (NavBar is hidden)
Select verse(s) and trigger the Bookmark action
BookmarkScreen loads (NavBar is hidden)
Press back on System Navigation
Android pops BookmarkScreen to show the ChapterScreen
Since BookmarkScreen showed the Navbar in the detach method, it's shown
when ChapterScreen is loaded again.

========

The ideas is that main content child fragments will individually decide
to either show or hide the main activity navigation component.

This change is necessary so that the Navbar doesn't show up at
unnecessary places.

Child fragments must show of hide it when they are attached.