UIX:ChapterScreen:Do not switch NavBar buttons, use in built buttons
Before:
When ChapterScreen is loaded, the MainNavBar showed a menu related to
actions that can be performed on the selected verses in the
ChapterScreen.

After:
When ChapterScreen is loaded, hide the MainNavBar
When a verse is selected, reveal a bottom bar containing possible
actions.

Other changes
Use Cards for title and actions container.
Verses will no longer be cards