ScreenSearch - UIX - Move actions to screen bottom when results are shown.
When results are present and shown hide the input field on top. show a
bottom button bar at the bottom. Initially the only button on the bottom
bar will be a reset button. But when the user selects verses in the
results list, new buttons will appear for each of the actions.