Fix - Config change does not update DailyVerse text formatting
The formatted text of the daily verse was cached earlier.
Then on config changes we matched conditions (reference + day of year +
default reference) and reused the previously cached formatted text.

This meant that when themes toggled or orientation changed, we still
reused the incorrect earlier cached formatted text.

To fix this
- We no longer cache the finished text
- We now cache the Book too
- Using the Book and Verse (already cached) we recreate the DailyVerse
text on each redraw of the view.