Create new global action that resets to home page on ErrorScreen load.
Depending on the flag passed in exitApp to the base method
showErrorScreen the correct navigation destination will be invoked.

If the App must exit from the error screen, then the entire app's
back stack will be popped.
Otherwise the stack will be popped up-till the HomeScreen - essentially
resetting the application to it's initial state.