Implement Preference action to select daily reminder notification time
Clicking the notification time preference entry will show a TimePicker
If the Positive button is selected, then the selected Hour & Minute is
saved in the SharedPreferences.

When the PreferenceScreen is loaded and displayed, we check if a valid
value for the key is present.
If yes, use it when displaying a TimePicker and also display the same
value in the locale format in the summary text of the record.