FIX - Cache and show John 3:!6 when the reference for the day fails
Earlier if the reference in the array failed, the method would return
without caching anything but it would display the Default Verse (John
3:16). This would mean that the bookmark and chapter actions failed
gracefully since there was no cached verse to act on.

So now, we will recursively call the same method with the default
reference to make sure that it is cached if the earlier reference fails.
Also in the method that formats and displays the verse, we check to see
if the target verse is the default reference, if yes, we use the
predefined styled text method to show it.